backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Django Backend"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
whitenoise = "^6.8.2"
redis = "^5.2.1"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pylint = "^3.0.3"
pytest = "^7.4.3"
pytest-django = "^4.7.0"
django-debug-toolbar = "^4.4.6"
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
